requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
idna==2.7
more-itertools==4.2.0
Pillow==5.2.0
pluggy==0.6.0
py==1.5.4
pytest==3.6.3
pytz==2018.5
PyYAML==3.13
rdoclient-py3==2.0.5
requests==2.19.1
six==1.11.0
urllib3==1.23
